fix: Fix ioctl(FIONREAD) type mismatch in tcp_status_stream()
The bytes_pending variable was declared as long (8 bytes on ARM64)
but ioctl(FIONREAD) writes an int (4 bytes). This could leave upper
bytes uninitialized with garbage values on ARM64.

This is the same bug that was fixed in tcp_read_until_condition() in
commit 38871d17.

+++ Common/source/tcpverbs.c
@@ -1522,7 +1522,7 @@ boolean tcp_status_stream(long stream_id, bigstring status_out, long *bytes_pend
     fd_set readset;
     struct timeval tv;
     int select_result;
-    long bytes_pending = 0;
+    int bytes_pending = 0;  /* Must be int for ioctl(FIONREAD) */
 
     log_debug(LOG_COMP_LANG, "tcp_status_stream: stream_id=%ld", stream_id);
